+++ tests/test-project/on_package_changed.py
@@ -1,0 +1,5 @@
+import json
+from sys import argv
+
+
+print(f"!!! PACKAGES CHANGED !!!{json.dumps(argv)}")