+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Poetry-Lock-Listener Changelog
+## 0.2.3
+### Fixed
+* fixed issues with poetry 2 compatibility
+### Internal
+* added component tests using docker
 ## 0.2.2
 ### Fixed
 * project now supports poetry 2.*